requirements.txt
Django==1.11.5
django-extensions==1.9.1
ipdb==0.10.3
ipython==6.2.0
Pillow==4.3.0
psycopg2==2.7.3.1
sorl-thumbnail==12.3
django-mptt==0.8.7